Retain messages after clearing
When a message is cleared, move it to a list of read messages rather
than deleting it for good. This alleviates the need to reload if you
want to bring back read messages.

When requesting archived messages, pass the asOf value in the
querystring so that previously-seen messages aren't resent.