Switch to environment variables for application configuration
Instead of using the node-config package, load configuration settings
from env.json if present. This file specifies values that are
otherwise expected to be available as environment variables.

Grunt will load this file directly and not populate or consider
process.env since the grunt tasks that need these kinds of values are
relevant in a dev environment.

In server.js the file will be loaded if it exists and process.env will
be populated. From then on server.js will look at process.env without
regard for how it was populated.

Grunt tasks for rsync and ssh have also been removed due to lack of
need.