Fix sent-headers error on /archive; switch response logging to finish event
The archive route was throwing a "Can't set headers after they are
sent" error, which was ended by taking out the next() call. But this
prevented the "end" log event from getting called (probably the wrong
setup from the start). Instead, use a single middleware instance for
logging-- log the request as soon as it arrives, and the response when
it has finished.