Handle OneDrive auth via server endpoints
The database backup script was previously set up manually, and
process for setting that up again wasn't obvious.

Instead, initiate the oauth flow for OneDrive via the server and store
the resulting tokens in the database. This is a first step toward
making the backup process internal to the server.