Migrate database backup script to new OneDrive API
OAuth will still be handled in the browser, but will not store tokens
in the database as previously planned. Doing so would make them harder
to access from the backup script. Using the previous approach of
writing a json file to the filesystem instead.

The backup script is otherwise migrated from the previous
skydrive/live.com API to the OneDrive API, which allows items to be
referenced by path instead of exclusively by ID.