Truncate message body when sending a browser notification
Chrome displays extra whitespace at the bottom of the notification
window if the body is longer than two lines. The icon area does not
expand to match the height of the body. Truncation avoids this and
keeps things brief. Other browsers don't need this limitation, but
it is applied to all.