Reorganize livereload env vars and simplify CSP header population
Pass a hostname as well as a port to grunt-contrib-watch, so that the
local machine hostname isn't used by default unexpectedly. Rename
NOTIFIER_DEV_HOST to NOTIFIER_LIVERELOAD_HOST since that's all it's
being used for. Populate the CSP header in a slightly less verbose
fashion, accounting for the livereload hostname and removing
unsafe-inline.