Reduce spurious browser notifications
Don't create browser notifications when the queue is being filled for
the first time. This comes into play with page reloads triggered by
livereload. In normal use the application is probably in focus
when the first fill occurs, so a notification isn't sent anyway.

Do send a notification if the browser is not focused and a new message
comes in (as before). Also fixed a typo when deciding whether an
existing message should be dropped (such as when coming back online
after a disconnect) that was causing all messages to be discarded.