Remove automatic reload on appcache update
The page reload triggered by an appcache update wasn't as useful as
intended. It would work better for very long-running browser sessions,
but for scenarios where you open and close the page throughout the day
there isn't as much payoff. You'll get the changes eventually.

On mobile, page refresh occurs in spite of appcache. So having an
additional reload there even if its legitimate is annoying.