Display a loading icon during connection setup
When the page initially loads, it is empty except for the application
message. This can make the page seem wrong or broken in certain cases,
such as when the page is shown in the window list of mobile Safari.

A loading icon is more interesting to look at than the previous
"Connecting..." application message, and is faded out pleasantly once
a connection has been established.