Use page header to display connection status only
The clear all link now appears in the footer, because I rarely ever
use it. This also simplifies the structure of the page header,
requiring one less template.

The connection status now only indicates the disconnected
state. Indicating the connected state shouldn't be necessary because
it is the normal state. This frees up more room for message display,
which is more important.