Get the connection icon to display at the correct size in Firefox
The icon was previously filling the bounds of the window instead of
displaying at a reasonable size, but only in Firefox. Firefox seemed
to ignore the elemnt selector in the external stylesheet.

Switching to a use element and moving the fill styling into the parent
selector fixed this. As an added bonus, the connection icon is handled
the same as the rest of the inlined svg files.

Safari was only willing to display the icon once the svg symbols were
moved to the top of the body tag rather than then bottom.