Get appcache working in Firefox
Firefox was unwilling to reload the appcache manifest. Whereas other
browsers would pick up changes after a second reload (once to trigger
a manifest request, a second for the newly downloaded files to be
used), Firefox was downloading once and then leaving the application a
cache orphan. The only way out was to clear the appcache from the
browser ('appcache clear' in the Firefox command bar).

Setting Cache-Control: no-cache private from the express static
middleware seems like the wrong thing to do, but it makes Firefox
behave as expected. The impact of having cacheable files that aren't
cached is mitigated by their being pulled from appcache. So it all
boils down to doing the wrong thing in terms of HTTP caching so that
the right thing happens in terms of applicationCache caching.