Apply focus style via css class, not via pseudo-selector
When the article tag of a message is focusable, clicking on the link
or close icons causes the message to come into focus and requires a
second click to achieve the original intent. The parent's focus is
taking precedence over a child element's click.

Using a class and removing tabindex from the article tag avoids this
while preserving the UI effect when using keyboard shortcuts.