Consolidate build and pre_build scripts
The build and pre_build scripts were originally separated to ensure
that npm install gets called with the custom-installed version of node
rather than the system version. That's actually not necessary.

Relying on a globally-installed version of grunt is also not
necessary. Switching to a local install triggered via an npm script
defined in package.json means one less thing to deal with during the
pre_build hook. When triggered via npm run-script, the local grunt
binary is used automatically.

There's still an unfortunate need to move package.json out of the way
to prevent the default OpenShift hook from invoking the system version
of node.