Simplify markup of status indicator in header
Reduce CSS hierarchy to only need one class to be toggled. This is an
effort to prevent a broken state that was once observed on iOS where
the status was in the disconnected state but the text shown was for
the connected state. This may have been a fluke though.