Don't show the time of disconnect
The time frequently ends up being the current time, rather than a point
in the past as was originally intended. A new session that fails to
connect for some reason will be offline as of the current time, which
isn't as meaningful as it would be for a long-running session. The
time is still otherwise available via the console.