Start dev server from an npm script rather than from grunt
Grunt was previously used to run a shell command which executed
nodemon. Nodemon provides automatic server restart when the source
file is modified.

Doing it this way is good in that build-related tasks are all handled
by grunt, but bad in that there is unnecessary overhead from having a
long-running grunt process. It also differs from how the production
server is started.

By using an npm script in package.json, the production and dev servers
are configured in the same location.