Build the UI earlier in the build lifecycle
If the application starts before the UI has been built, files may be
missing that cause the application to throw an error. This occurred
with the express favicon middleware, which tried to load a favicon
file that had not yet been created.

It's also more intuitive if building occurs from the build hook.