Correct uneforced guard against too-soon message list refetch
There's a check within the MessageList service to prevent fetch from
being called more than once per second, but it wasn't returning
properly and there was a redundant call to the archive endpoint.

In the status bar, only set the message status for offline or
disconnect. If it is set for all states, a connection event can
clobber a previously set tally summary because a previous fetch
commpleted before the socket connection completed.