Revise format of response from archive endpoint for flexibility
The archive endpoint previously returned a list of messages. Now it
returns an object with a messages key containing the list. This
provides room to send back additional values. The limit key is now
being returned indicating what the count parameter was changed to
during min/max validation. For example, a request for /archive/999
now returns {limit: 100, messages: [...]}.