Retract unread messages based on incoming local id
When a new message arrives with the same local id as one or more
existing unread messages, retract the existing messages. This
alleviates the need to clean up an earlier instance of a repeating
message, such as a daily reminder. It can be avoided by assigning
unique local ids, or by not assigning a local id at all.