Listen for error event during appcache manifest check
If the server is unavailable due to a 503 or otherwise, trigger a page
reload after 5 seconds via the appcache reload directive. This won't
handle the case where the cache manifest request fails but then the
server comes back online in time for the archive request to
succeed. It's only considering what happens from the first point of
contact with the server.