Prevent CSP violation in Safari related to browser extensions
Safari 9 was throwing a CSP violation message related to the
script-src directive. The culprit was uBlock. Even when deactivated
for the site, the CSP violation would occur. There was no violation
when the extension was disabled.

Allowing unsafe-inline on script-src is not particularly desirable,
but at least it's restricted to Safari only.