Remove backgroundcolor from message options
There wasn't a strong reason for the clear icon to have a separate
background color than the rest of the message, and in Safari there was
an elusive 1px gap at the bottom that was annoying. Not having a
background color makes the icon feel like an inherent part of the
message instead of UI chrome.