Make access logging file-based, and query logging off by default
Express access logs now go to a file in combined format, as is typical
of a web application. Sequelize logs queries to the console by
default, but rather than also send them to a file, provide an on/off
flag in the configuration to easily turn them on when needed.