Add support for setting a base url
The application previously expected to run from the root URL. This
causes problems if you want to run from a subdirectory, either
directly or through a proxy. Use a base tag to capture this on the
client side, and account for it in the Grunt build process and the
server. Use the root URL as a default.