Use an ejs template to render the homepage
This change is motivated by the need for base url flexibility. With a
static homepage templated through Grunt, changing the base URL would
require a rebuild. Dropping this setup makes for a simpler build
process as well.

Nodemon has also been reconfigured so that it picks up changes to ejs
templates as well as modules.