Generate the appcache manifest from EJS template
Grunt-based manifest generation used absolute URLs, creating
complications with the base url during deployment. The files listed in
the manifest rarely changes, so the pattern-based flexibility provided
by grunt-appcache wasn't a benefit.

The root-level build.txt is introduced as a manifest unique
identifer. It allows the manifest to reflect the recent build date,
which is otherwise only known to Grunt.