Do away with the external-link icon
Placing the badge was awkward on mobile and the badge distracted from
the icon on both mobile and desktop. Since the message othewise
displays the link domain if the message is clickable, try just going
by that alone to distinguish clickable from non-clickable messages.

In day-to-day use, clicking on a message is less of a UI driven
behavior than a learned behavior.