Don't clear the status message during connection:change
This line was the cause of the status bar being blank on initial
load. It was being populated, but then cleared immediately after. For
reasons unknown, the connection change seems to be happening after the
queue change.

Clearing the message otherwise unnecessar, because when returning to a
connected state there will be a queue change event right after anyway.