Allow data: in CSP connect-src for Safari
Message bodies are allowed to contain a limited set of HTML tags. The
ng-bind-html Angular directive is used so that this markup is
preserved if it is present. This causes angular to render the string
as a data URL, which Safari considers a violation of the CSP
header. Although Chrome and Firefox do not.

Add data: to connect-src to prevent this. The change doesn't take
effect immediately, because Safari seems to cache the CSP header's
value.