Maintain 100% height in a more reliable way
Rather than cascading down 100% from the HTML tag to the main element,
use min-height specified in vh on the main element. This achieves the
desired effect of maximizing the height when there is less one screen
of content, and exceeding it when there is.

Setting justify-content back to flex-start prevents weird behavior
where the content is cut off due to being top-shifted.