Refactor web worker to a class
Separate the event handling of the worker with the management of the
EventSource instance. This version includes dormant methods in the
receiver class for client-side retry, which don't seem to be
worthwhile compared to the built-in retry behavior of
EventSource. Keeping them in for future reference but will prune in a
subsequent commit.