Move server to a top-level directory, add webpack
Following the earlier relocation of the worker script to the top
level, the server and app now follow suit. These three directories are
the main pieces of the application.

The introduction of webpack is a change-of-course from an earlier
start with browserify. Forced by the need to make modules produced by
typescript palatable to the browser, webpack's bundle ends up
delivering a relatively tidy end result.