Calculate appcache manifest build hash on-the-fly
This was previously handled by a build.txt file emitted by Grunt at
the end of a build, and included by the appcache ejs
template. Carrying that approach over to webpack could work, but would
add extra complexity to the webpack config and generally be a
hassle. Handling this dynamically from the server is more convenient.