Repair broken handling of pushbullet identifier
The id returned from pushbullet wasn't being stored on the message due
to confusion about the value of the message. When the update occurs,
the value is a plain JS object and not a full model instance, so there
is no id value.

Use publicId for the where clause of the update instead.