Pull dev-environment build steps back into run scripts
The previous intent was to run everything through Make, and not use
NPM scripts. This didn't work out well with tmux. Killing the tmux
session would leave behind any of the processes that Make had started,
such as the devserver and the webpack watches.

Running a server process isn't really something that Make is geared
towards. On the other hand, npm scripts are not able to express
dependencies as well as Make can.

Instead, use both: scripts for managing long-running processes, but
not shell-related tasks. Make targets for shell-reltaed tasks, but not
long-running processes.