Adjust syntax for ng-repeat used to display shortcuts
The shortcut map was previously an object, but now it is an
array. Angular needs to be told how to track the items in the array
and to not display undefined items (the gaps in the array where no
value has been defined).

The shortcut map was switched to an array because the indexes are
numeric. They become strings when used as object keys, which doesn't
make much sense here.