Scale back the handling of appcache updates
Only reload the page in response to an updateready event. Ignore
errors, and skip any console logging.

The previous approach of keep-reloading-until-successful was nice in
theory, but annoying in practice. If the server is unreachable, a
reload loop can occur. Outside of development, appcache updates happen
infrequently and it's not the end of the world if the latest version
isn't immediately picked up.