Refactor server to not use node-ts
Compiling ts files on-the-fly via ts-node is convenient in dev, but
awkward in prod. In prod, typescript compilation should only have to
happen once. Once compiled, the TS files are no longer needed.

The build folder now serves as the application root. The public folder
resides there along with a JS-version of the server, managed
respectively by webpack and tsc configs. Annoyingly, the ejs views
move there as well since they are not managed by tsc and would
otherwise be left out.

Type definitions at the project root also move into the server, as was
done for the app and worker.