Perform app minification outside of webpack
The version of uglifyjs provided by webpack produces a minified
version of app.js that throws a surprise runtime error related to a
custom angular filter. The webpack documentation describes how to use
a newer version of the uglifyjs plugin, but doing so involves a plugin
whose lack of type definitions makes Typescript balk, since it is
otherwise running in strict mode. Stubbing in a type definition for
uglifyjs-webpack-plugin was not straightforward.

Invoking uglifyjs directly from the Makefile is far more
straightforward.