Add support packages for postgres
The sqlite3 package that is expected by Sequelize is a hassle to work
with since it needs to be compiled via gyp. Having this compilation
step complicates CI when there are multiple architectures involved
(production is arm, CI server is not). It also makes for a
ridiculously long build time if the machine doing the compile is on
the lower end. It also tends to be flaky and not work for no apparent
reason.

Although SQLite is otherwise a viable data source for this
application, the Node drivers for Postgres don't involve this kind of
compilation which makes things a lot easier.