Remove NOTIFIER_LOG_QUERIES
This was meant to be a toggle for controlling Sequelize's default
behavior of printing SQL queries to the console. Aside from being
inadvertently deactivated, this feature was rarely used and the
verbosity of queries that weren't otherwise of interest was
annoying.

Query logging can still be enabled from the main server file if
needed.