Invert style definition of page jacket offline state
Rather than defaulting to the offline state and activating the online
state based on the presence of a class, do it the other way
around. This avoids a problem with the page seemingly getting stuck in
the disconnected state after reconnect because the class ends up being
"online" rather than "connected".