Take badges out of appcache
Downloading all the badges via the appcache manifest isn't worth
doing. If the client is online to receive a message, it can also
request a badge. There is also a slight waste of time from downloading
badges that may never get used. But more importantly if the number of
badges increases significantly things would get out of hand.