Allow messages to specify a custom badge image
The server can be optionally configured to reference badge images
hosted externally. If an incoming message specifies a badge value, it
will be treated as a filename on that host. If no badge is specified,
the default for the group is displayed instead.

Badges are restricted to a single domain and meant to be sourced from
this one location only. This is reflected in the img-src property of
the Content Security Policy header.