Enforce message expiration on the server
When a message expired, the server took no action and instead relied
on the browser client to send a retraction. This didn't account for
integration with external services, who never heard about the
retraction unless a browser client was present to initiate it.

Now the server will track expiring messages internally and be more
proactive about retractions across all services.

There is redundancy with the existing functionality in the browser
client, but that has been left as-is so that expiration can be carried
out if the client is offline.