Maintain message summary after reconnect and drop per-group counts
If there was a disconnect, the message summary was being overwritten
by the most recent connection event. So you'd see "connected" but not
the message count, which is more meaningful.

The tally of messages in each group wasn't all that useful because the
group names aren't always meaningful. They also aren't equally
actionable.