Revert to webpack 3
Upgrading to webpack caused both the app and worker builds to
fail. The worker was fixable with a little effort due to its
simplicity, but the app is less straightforward due to use of the
extract text plugin.

Also, and worse, the upgrade prompts for installation of a separate
package for the CLI which in turn brings in 330 additional
packages. That's insane.