Add dom lib to worker tsconfig
This gets the worker build working again. It was failing because it
couldn't find type information for the EventSource object. Although
the worker tsconfig pulls in the TypeScript WebWorker type lib, this
lib no longer accounts for EventSource. So even though the dom
TypeScript lib is not otherwise needed by the worker, its inclusion in
the tsconfig is a means to an end.