Remove SVG symbols from index.html
Handle the close and no-messages icons the same way that badge icons
are handled: as CSS background images. Although this makes styling
more complicated, it reduces the amount of markup rendered by mithril
and increases overall consistency. It also avoids a "Attribute without
value" error when doing a parcel build.