Reorganize dev setup and relocate public folder
The previous setup was awkward due to the server running transpiled
Typescript files, which resided in a separate build directory shared
with the UI files. The current setup is to instead transpile
on-the-fly. Nodemon is replaced with ts-node-dev. The public directory
becomes part of the server's filesystem tree and parcel writes it
files there. The build directory is no longer used.