Discontinue use of session storage.
Reduce complexity by not having to keep session storage up-to-date
with the in-memory map holding the message list. It was originally
done that way to make the message list appear more quickly, but that
benefit only holds while the UI is initially rendered. From then on,
session storage isn't useful.