Rework server config for Node v14.
Remove inheritance from the server tsconfig.json so that it is more
explicit, and reduce the contents to the bare minimum to get the
server running without error. This forces changes to how modules are
being imported for reasons that are not entirely clear. It's also
less-than-clear why trying to use ES modules seemed to make things
worse.